[Spring 지하철 노선도 관리 - 3단계] 아마찌(신지혜) 미션 제출합니다. (#162)
* refactor(test): request-body를 json으로 파싱하도록 변경

* refactor(test): 중복되는 테스트 메서드 추출

* refactor(line-service): 서비스 레이어 추가

* refactor(station-service): 서비스 레이어 추가

* style(): 리포맷팅 적용

* feat: 3단계 DB설계

* refactor(): 도메인 위주의 패키지 분리

* feat(SectionDao): Db 구간 저장 로직  구현

* feat:  POST /lines 노선 추가시 구간 정보도 함께 등록

* feat(): 구간 추가 서비스 로직 정리

* refactor(SectionEntity.clss):  DB로직을 위한 Entity객체 생성

* refactor(LineEntity.class): DB로직을 위한 Entity객체 생성

* refactor(LineEntity.class): Entity 속성 변경

* refactor(): 패키지 구조 변경

* refactor(): controller-advice 추가

* refactor(): spring validation을 이용하여 dto 값 검증하도록 변경

* refactor(line-response): dto 내부에서 필요한 값을 세팅하도록 변경

* refactor(line): line 도메인 하나로 관리하도록 변경

* feat(sections): 구간들을 관리하는 일급 컬렉션 추가

* feat(sections): 노선에 구간을 등록할 수 있는지 확인하는 기능 추가

* refactor(line): section 확인하는 기능 이동

* refactor(): 패키지 구조 변경 및 병합하다 누락된 파일들 추가

* refactor(): 불필요한 dto 제거

* feat(): 노선 조회 시 구간에 포함된 역 목록 응답 기능 구현

* refactor(): 불필요한 클래스 제거

* refactor(): service에서 repository에 의존하도록 변경

* feat(): 구간 추가 API 구현

* refactor(): dao-test에서 dummy.sql 사용함을 명시

* feat(): 구간 추가 기능 구현

* test(): 도메인 테스트 추가

* test(section-dao): 부족한 테스트 코드 추가

* refactor(dao): clear 메서드 제거

* feat(station): 지하철 역 제거 예외처리 기능 추가

* test(station): 구간에 등록된 역 제거 예외 테스트 구현

* feat(line): 해당 역을 포함한 구간을 반환하는 로직 추가

* fix(): 종점에 구간 추가 되지 않았던 오류 수정

* refactor(): 외래키 제약조건 수정 및 노선 삭제 기능 구현

* refactor(): 패키지 이동

* test(): 구간 추가 삭제 확인

* refactor(): 불필요한 메서드 제거

* refactor(): dto 관련 exception handler 변경

* refactor(): 서비스 레이어에서 응답 dto를 반환하도록 변경

* refactor(): 테스트에서 dto 빠진 부분 수정

* refactor(): 서비스에서 도메인으로 변경하도록 변경

* refactor(): line-repository에서 section-dao에 접근하도록 변경

* refactor(): line-repository에서 section-dao에 접근하도록 변경

* refactor(): 전략패턴을 이용하여 중복되는 구간 추가 로직 분리

* style(): 인텔리제이 리포맷팅 적용

+++ src/main/java/wooteco/subway/station/service/StationService.java
@@ -1,38 +1,46 @@
 package wooteco.subway.station.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.line.dao.SectionDao;
+import wooteco.subway.line.domain.Section;
 import wooteco.subway.station.dao.StationDao;
-import wooteco.subway.station.domain.Station;
+import wooteco.subway.station.dto.StationDto;
 
 import java.util.List;
 
 @Service
 public class StationService {
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
-    public Station save(Station station) {
-        return stationDao.save(station);
+    @Transactional
+    public StationDto save(StationDto stationRequest) {
+        validateDuplicate(stationRequest.getName());
+        return StationDto.toDto(stationDao.save(StationDto.toStation(stationRequest)));
     }
 
-    public List<Station> findAll() {
-        return stationDao.findAll();
+    private void validateDuplicate(final String name) {
+        if (stationDao.findByName(name).isPresent()) {
+            throw new IllegalStateException("[ERROR] 이미 존재하는 역입니다.");
+        }
     }
 
-    public Station findById(Long id) {
-        return stationDao.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("[ERROR] 존재하지 않는 역입니다."));
+    public List<StationDto> findAll() {
+        return StationDto.toDtos(stationDao.findAll());
     }
 
-    public Station findByName(String name) {
-        return stationDao.findByName(name).orElseThrow(() ->
-                new IllegalArgumentException("[ERROR] 존재하지 않는 역입니다."));
-    }
-
+    @Transactional
     public void delete(Long id) {
+        List<Section> sections = sectionDao.findByStationId(id);
+        if (!sections.isEmpty()) {
+            throw new IllegalStateException("[ERROR] 구간에 역이 등록되어 삭제할 수 없습니다.");
+        }
         stationDao.delete(id);
     }
 }
